Report errors to client when sending the RETR data
Before this change we ignored the errors returned from copying the
data to the socket.

This meant that if an error occurred the client would hang
indefinitely - the server thinking it had finished and the client
waiting for a response from the server.

After this change the server responds with a 551 message if the stream
breaks during the transfer.  This allows the client to pick up the
error and carry on.